adding io apic

+++ kernel/arch_x86_64/sources/arch_smp.c
@@ -328,7 +328,7 @@ INIT_TEXT int get_gsi_for_irq(int irq) {
 // 8259A PIC 中断触发条件是 edge-triggered high pin polarity（信息来自 IBM PC/AT Technical Reference）
 
 INIT_TEXT int get_gsi_trigmode(int gsi) {
-    if (gsi <= g_gsi_max) {
+    if ((uint32_t)gsi <= g_gsi_max) {
         switch (TRIGMODE_MASK & g_gsi_flags[gsi]) {
         case TRIGMODE_EDGE:  return 1;  // edge
         case TRIGMODE_LEVEL: return 0;  // level
@@ -342,7 +342,7 @@ INIT_TEXT int get_gsi_trigmode(int gsi) {
 }
 
 INIT_TEXT int get_gsi_polarity(int gsi) {
-    if (gsi <= g_gsi_max) {
+    if ((uint32_t)gsi <= g_gsi_max) {
         switch (POLARITY_MASK & g_gsi_flags[gsi]) {
         case POLARITY_HIGH: return 1;   // high
         case POLARITY_LOW:  return 0;   // low
